refactor(routing): retire reflection-era dead code after #1041 (#1042)
Remove machinery the #1041 rotation landing left unreferenced, keeping every
render byte-identical:

- Delete offsets._is_over_top_right_entry (no production caller after #1041;
  only the seam-orientation corpus oracle used it). Relocate it into
  tests/test_seam_orientation.py, which keeps it as ground truth.
- Delete the unread _RoutingCtx.tb_right_entry field and its computation in
  context.py (tb_positive_fan_sections recomputes tb_right_entry_sections
  directly; the context field had no reader).
- Prune dangling references to the already-deleted _reverse_tb_right_entry_offsets
  in inter_section_handlers.py and offsets.py, pointing instead at the live
  _reorder_reconvergence order-half.

Regenerate the routing gate-coverage baseline + matrix for the removed gate arms.

Residuals kept and documented (the seam classifier cannot derive them
coordinate-free, so they ride their own reversal pass, not the per-section lane
sign): _reverse_around_below_left_entry_offsets (horizontal around-below LEFT
half-turn) and _reverse_near_vertical_junction_right_entry_offsets (RIGHT entry
whose transpose turns on junction pixel overhang). The RL (local_max - priority)
offset reflection gated by _stores_reflected also stays: it is load-bearing for
the byte-identical RL renders.
