+++ src/nf_metro/layout/routing/offsets.py
@@ -7,7 +7,6 @@ from collections.abc import Sequence
 from dataclasses import dataclass, field
 
 from nf_metro.layout.constants import (
-    COORD_TOLERANCE,
     COORD_TOLERANCE_FINE,
     OFFSET_STEP,
     SAME_Y_TOLERANCE,
@@ -2126,43 +2125,6 @@ def compute_station_offsets(
     return ctx.offsets
 
 
-def _is_over_top_right_entry(
-    graph: MetroGraph, port: Port, tb_sections: set[str]
-) -> bool:
-    """Whether *port* is a RIGHT entry reached by an over-the-top loop.
-
-    Matches the dispatch of ``_route_right_entry_over_top``: a RIGHT entry on a
-    TB section fed by an exit port in the SAME grid row, an ADJACENT column, and
-    to the port's LEFT.  That feed loops over the section's top and approaches
-    from the right -- a U-turn that transposes the bundle.  A right entry fed
-    from the right (a fold) or across columns (a bypass) keeps its order and is
-    excluded.
-    """
-    if not (port.is_entry and port.side == PortSide.RIGHT):
-        return False
-    if port.section_id not in tb_sections:
-        return False
-    psec = graph.sections.get(port.section_id)
-    pst = graph.stations.get(port.id)
-    if psec is None or pst is None:
-        return False
-    for edge in graph.edges_to(port.id):
-        src = graph.stations.get(edge.source)
-        src_port = graph.ports.get(edge.source)
-        if not (src and src_port and not src_port.is_entry):
-            continue
-        ssec = graph.sections.get(src.section_id) if src.section_id else None
-        if ssec is None:
-            continue
-        if (
-            ssec.grid_row == psec.grid_row
-            and abs(ssec.grid_col - psec.grid_col) <= 1
-            and src.x < pst.x - COORD_TOLERANCE
-        ):
-            return True
-    return False
-
-
 def _reverse_offsets_from_roots(ctx: _OffsetCtx, roots: set[str]) -> None:
     """Reverse the per-line order of *roots* and their DAG-downstream sections.
 
@@ -2202,13 +2164,16 @@ def _reverse_offsets_from_roots(ctx: _OffsetCtx, roots: set[str]) -> None:
 def _reverse_around_below_left_entry_offsets(ctx: _OffsetCtx) -> None:
     """Reverse the line order of sections entered through an around-below LEFT port.
 
-    The mirror of :func:`_reverse_tb_right_entry_offsets` for the horizontal
-    idiom: a reverse-flow bypass leaving a LEFT exit, dropping below every box,
-    and rising into a far-side LEFT entry is a half-turn that transposes the
-    bundle end-to-end (see ``_route_left_exit_around_below_left_entry``).  The
-    section therefore receives its lines in the opposite order to the source, so
-    its entry port and internal trunk carry the reversed order for the rise into
-    the port and the run out of it to stay straight.
+    A reverse-flow bypass leaving a LEFT exit, dropping below every box, and
+    rising into a far-side LEFT entry is a half-turn that transposes the bundle
+    end-to-end (see ``_route_left_exit_around_below_left_entry``).  The section
+    therefore receives its lines in the opposite order to the source, so its
+    entry port and internal trunk carry the reversed order for the rise into the
+    port and the run out of it to stay straight.
+
+    A horizontal idiom outside the vertical-flow rotation and the seam-orientation
+    reorder, so it keeps its own reversal pass rather than riding the per-section
+    lane sign.
     """
     graph = ctx.graph
     _reverse_offsets_from_roots(
@@ -2230,6 +2195,10 @@ def _reverse_near_vertical_junction_right_entry_offsets(ctx: _OffsetCtx) -> None
     into the port's lateral order, so the section receives its lines in the
     opposite order to the junction; it carries the reversed order so the drop and
     the run out of the port stay straight and the turn nests concentrically.
+
+    Whether the drop transposes turns on the junction's pixel overhang, not on
+    port sides or grid rows, so the seam-orientation classifier cannot derive it
+    coordinate-free; this pass stays as a coordinate-aware residual.
     """
     graph = ctx.graph
     _reverse_offsets_from_roots(
