+++ tests/data/routing_gate_coverage_baseline.json
@@ -206,13 +206,10 @@
   "normalize.py::if sg[0].sign_x != t.sign_x:::#1",
   "normalize.py::if trunk_ch is None:::#1",
   "normalize.py::if trunk_rp is None or trunk_src_st is None:::#1",
-  "offsets.py::and abs(ssec.grid_col - psec.grid_col) <= 1::#1",
-  "offsets.py::and src.x < pst.x - COORD_TOLERANCE::#1",
   "offsets.py::elif feeding_exit != edge.source:::#1",
   "offsets.py::for _ in range(max_iterations):::#1",
   "offsets.py::for edge in graph.edges_to(jid):::#2",
   "offsets.py::for edge in graph.edges_to(pid):::#2",
-  "offsets.py::for edge in graph.edges_to(port.id):::#1",
   "offsets.py::for lid, off in desired.items():::#1",
   "offsets.py::for lid, off in desired.items():::#2",
   "offsets.py::for other in lines:::#1",
@@ -246,8 +243,6 @@
   "offsets.py::if max_steps <= 0:::#2",
   "offsets.py::if neighbour is None:::#1",
   "offsets.py::if new_order is None:::#2",
-  "offsets.py::if not (port.is_entry and port.side == PortSide.RIGHT):::#1",
-  "offsets.py::if not (src and src_port and not src_port.is_entry):::#1",
   "offsets.py::if not all(lid in feeder_off for lid in continuing):::#1",
   "offsets.py::if not all(lid in feeder_off for lid in primary_lines):::#1",
   "offsets.py::if not any(::#1",
@@ -262,9 +257,7 @@
   "offsets.py::if ordered != expected:::#1",
   "offsets.py::if other_lid == cur_lid:::#1",
   "offsets.py::if paired_off is not None:::#1",
-  "offsets.py::if port.section_id not in tb_sections:::#1",
   "offsets.py::if port_off is not None:::#1",
-  "offsets.py::if psec is None or pst is None:::#1",
   "offsets.py::if reverse:::#3",
   "offsets.py::if seam is None:::#1",
   "offsets.py::if sec_a is None and id_a in graph.ports:::#1",
@@ -280,7 +273,6 @@
   "offsets.py::if src_port.section_id == feeder_id:::#1",
   "offsets.py::if src_st and not src_st.is_port:::#1",
   "offsets.py::if src_st and not src_st.is_port:::#2",
-  "offsets.py::if ssec is None:::#1",
   "offsets.py::if succ not in affected:::#1",
   "offsets.py::if swap_lid is None:::#1",
   "offsets.py::if tgt_col is None:::#1",
@@ -288,7 +280,6 @@
   "offsets.py::if trunk_feeder_id is not None:::#2",
   "offsets.py::or port.section_id in ctx.reversed_sections::#1",
   "offsets.py::sec is None::#1",
-  "offsets.py::ssec.grid_row == psec.grid_row::#1",
   "offsets.py::while cur is not None and cur not in seen:::#1",
   "offsets.py::while stack:::#1",
   "perp.py::if not indices:::#1",
